Add PUP::er::pup_buffer methods to allow flexible pupping/unpupping of buffers
The pup_buffer methods allow users to invoke a 'deallocate' method
to free the passed buffer, after packing as bytes and similarly invoke
an 'allocate' method to allocate a new buffer and assign it to the
passed buffer before unpacking as bytes. There is a default version,
where 'free' and 'malloc' will be called when the user doesn't pass
allocate and deallocate methods. The pup_buffer methods have the
following signature:

  void pup_buffer(T *&a, size_t nItems);

  void pup_buffer(T *&a, size_t nItems,
                        std::function<void *(size_t)> allocate,
                        std::function<void (void *)> deallocate);

This API was ideated by Evan Ramos from Charmworks, Inc.